Moved class to file
